pyproject.toml
[project]
name = "transporter_networks"
version = "0.0.0"
description = "Implementation of transporter networks."
authors = [
	{ name = "Peter David Fagan", email = "peterdavidfagan@gmail.com" }
]
readme = "README.md"

[tool.poetry]
name = "transporter_networks"
version = "0.0.0"
description = "Implementation of transporter networks."
license = "Apache-2.0"
authors = [
	"Peter David Fagan <peterdavidfagan@gmail.com>"
]
readme = "README.md"

[tool.setuptools]
py-modules=["transporter_networks"]

[tool.poetry.dependencies]
python = "3.10.6"
numpy = "^1.16.0"
pillow = "10.0.0"
jaxlib = {extras=["cuda12_pip"], url = "https://storage.googleapis.com/jax-releases/cuda12/jaxlib-0.4.23+cuda12.cudnn89-cp310-cp310-manylinux2014_x86_64.whl"}
matplotlib = "^3.7.2"
hydra-core = "^1.3.2"
pytest = "^7.4.2"
lark = "^1.1.7"
inflect = "^7.0.0"
einops = "^0.7.0"
clu = {path="./transporter_networks/CommonLoopUtils", develop=true}


[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120

